Provided answers to Part II
Answered all of the questions in Part II, but honestly I'm feeling like
the results I'm getting are kind of strange and not what I anticipated
for this project. It could potentially be a trick to make sure I
understand the practical applications, but after running through the
project completely I want to check all my answers again in case I
misunderstood something.