Proofread, added html version, submitted
A couple minor proofreading tweaks, downloaded and added to the folder
an html version of the notebook for readability. After going through
the notebook again I didn't find a reason to chance my calculations
from the first time through, and I only hope they're correct based on
the strange conclusions at the beginning that I know are correct that
the new page had a very slightly lower conversion rate than the old
page.